Moved Dashboard
Moved the Dashboard page into it's own folder.
