+++ betteradmin/src/Pages/Account/Dashboard/Dashboard.css
@@ -1,0 +1,11 @@
+.Hold-component {
+  border: 1px solid red;
+  }
+.Underviser-component {
+    border: 1px solid yellow;
+    padding: 10px;
+  }
+ .Lektier-component {
+    border: 1px solid green;
+    padding: 10px;
+  }
