+++ betteradmin/src/Pages/Account/Dashboard/Dashboard.jsx
@@ -1,0 +1,27 @@
+import './Dashboard.css';
+
+export default function Dashboard() {
+    return (
+        <>
+            <div className='Hold-component'>
+                <h1>Dine Hold</h1>
+                Klaver
+                12/12/12 
+            </div>
+            <div className='Underviser-component'>
+                <h1>Undervisere</h1>
+                Anders Andersen
+                Klaver Underviser
+                Start en chat
+            </div>
+            <div className='Lektier-component'>
+                <h1>Lektier</h1>
+                Til den 12/12/12
+                lær jule sange
+                19/12/12 
+                Jule event
+            </div>
+        </>
+    );
+  }
+
