--- betteradmin/src/Pages/Dashboard.css
@@ -1,12 +1,0 @@
-.Hold-component {
-    border: 1px solid red;
-    padding: 10px;
-  }
-.Underviser-component {
-    border: 1px solid yellow;
-    padding: 10px;
-  }
- .Lektier-component {
-    border: 1px solid green;
-    padding: 10px;
-  }
